+++ src/stats/matrix_adjust.cpp
@@ -1244,8 +1244,8 @@ vector<int> CompositionMatrixAdjust(int query_len, int target_len, const double*
         joint_probs,
         background_freqs);
     if (r != 0) {
-        for (int i = 0; i < TRUE_AA; ++i)
-            for (int j = 0; j < TRUE_AA; ++j)
+        for (size_t i = 0; i < TRUE_AA; ++i)
+            for (size_t j = 0; j < TRUE_AA; ++j)
                 v[i * 20 + j] = score_matrix(i, j) * scale;
         //throw std::runtime_error("Error computing composition matrix adjust.");
     }
